// src/lib/api/models/NoChangeOperation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NoChangeOperation = {
    type: NoChangeOperation.type;
    /**
     * Id of the existing image
     */
    imageId: string;
};
export namespace NoChangeOperation {
    export enum type {
        NOCHANGE = 'nochange',
    }
}